+++ tests/analysis/normalization/__init__.py
@@ -1,0 +1,5 @@
+"""
+Tests for the normalization modules
+
+Copyright © 2024 Pixelgen Technologies AB.
+"""